pyproject.toml
[project]
name = "img-archive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow-jxl-plugin>=1.3.3",
]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]
